Version: 0.0.27

+++ startApi.sh
@@ -3,5 +3,7 @@ git fetch origin
 git reset --hard origin/main
 yarn
 ./startDB.sh
-cp .env.prod .env
+if ! test -f .env ; then 
+    cp .env.prod .env
+fi
 HTTPS=true  node -r newrelic ./build/index.js
